+++ ydotoold.c
@@ -35,7 +35,7 @@ void * client_handler(void * fdp) {
     int fd = *(int *)fdp;
 
 	for (;;) {
-		int rc = recv(fd, &buf, sizeof(buf), MSG_WAITALL);
+		ssize_t rc = recv(fd, &buf, sizeof(buf), MSG_WAITALL);
 
 		if (rc == sizeof(buf)) {
 			uinput_emit(buf.type, buf.code, buf.value);
